upload image

+++ DoAnCShap/Frm_NhanVien.cs
@@ -26,7 +26,6 @@ namespace DoAnCShap
         }
         String TenHinh = "";
         String DuongDan = "";
-        String DuongDanFolderHinh = @"C:\Users\Nguyen Khanh\source\repos\DATN\DoAnCShap\Image";
         MD5 md = MD5.Create();
         NhanVien_BUS bus = new NhanVien_BUS();
         ChucVu_BUS buss = new ChucVu_BUS();
@@ -49,6 +48,21 @@ namespace DoAnCShap
             cboTrangThai.Enabled = b1;
         }
       
+        public void Clear()
+        {
+            txtMaNV.Clear();
+            cboChucVu.Text = "";
+            txtHinhNhanVien.Clear();
+            txtUserName.Clear();
+            txtPassWord.Clear();
+            txtTenNV.Clear();
+            cboGioiTinh.Text = "";
+            txtEmail.Clear();
+            txtSDT.Clear();
+            txtCMND.Clear();
+            txtDiaChi.Clear();
+            cboTrangThai.Text = "";
+        }
         public void xulychucnang(Boolean b1, Boolean b2, Boolean b3)
         {
             btnThem.Enabled = b1;
@@ -239,24 +253,23 @@ namespace DoAnCShap
             txtSDT.Text = row.Cells[6].Value.ToString();
             txtCMND.Text = row.Cells[7].Value.ToString();
             txtDiaChi.Text = row.Cells[8].Value.ToString();
-            //string[] b = row.Cells[9].Value.ToString().Split(';');
-            //pictureBox1.Controls.Clear();
-            //int n;
-            //if (b.Length == 1)
-            //    n = b.Length;
-            //else
-            //    n = b.Length - 1;
-            //for (int i = 0; i < n; i++)
-            //{
-            //    PictureBox p = new PictureBox();
-            //    Size s = new Size(150, 150);
-            //    p.Size = s;
-            //    pictureBox1.Controls.Add(p);
-            //    Bitmap a = new Bitmap(DuongDanFolderHinh + "\\" + b[i]);
-            //    p.Image = a;
-            //    p.SizeMode = PictureBoxSizeMode.StretchImage;
-            //}
-            pictureBox1.Image = Image.FromFile("C:\\Users\\Nguyen Khanh\\source\\repos\\DATN\\DoAnCShap\\Image\\"+txtHinhNhanVien.Text);
+            string[] b = row.Cells[9].Value.ToString().Split(';');
+            pictureBox1.Controls.Clear();
+            int n;
+            if (b.Length == 1)
+                n = b.Length;
+            else
+                n = b.Length - 1;
+            for (int i = 0; i < n; i++)
+            {
+                PictureBox p = new PictureBox();
+                Size s = new Size(150, 150);
+                p.Size = s;
+                pictureBox1.Controls.Add(p);
+                Bitmap a = new Bitmap(DuongDan + "" + b[i]);
+                p.Image = a;
+                p.SizeMode = PictureBoxSizeMode.StretchImage;
+            }
             txtHinhNhanVien.Text = row.Cells[9].Value.ToString();
             txtUserName.Text = row.Cells[10].Value.ToString();
             txtPassWord.Text = row.Cells[11].Value.ToString();
@@ -273,16 +286,39 @@ namespace DoAnCShap
 
         private void btnXoa_Click(object sender, EventArgs e)
         {
+           DialogResult KQ = MessageBox.Show("Bạn có muốn xóa hay không ?", "Thông Báo !!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (KQ == DialogResult.Yes)
+            {
             nv.MaNV = txtMaNV.Text;
             bus.DeleteData(nv);
             MessageBox.Show("Xóa Nhân Viên Thành Công");
             Display();
+            }
+            else
+            {
+                MessageBox.Show("Hủy Xóa !");
+            }
         }
 
         private void txtSearch_TextChanged(object sender, EventArgs e)
         {
             string Condition = txtSearch.Text;
             HienThiSearch(Condition);
+        }
+
+        private void btnHuy_Click(object sender, EventArgs e)
+        {
+            DialogResult KQ = MessageBox.Show("Bạn có muốn hủy hay không ?", "Thông Báo !!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (KQ == DialogResult.Yes)
+            {
+                Clear();
+                xulychucnang(true, false, false);
+                xulytextbox(false, true);
+            }
+            else
+            {
+
+            }    
         }
     }
 }
